### Description
During task submission to cluster manager node, cluster manager node performs de-duplication checks to validate there are not any duplicate task already in queue. If there are duplicate task in queue, it will throw exception to new duplicate task submission.

#### Current logic
Cluster manager node has `tasksPerBatchingKey` map, which stores all the tasks which are batched against a task executor. In current de-dup logic, cluster manager node fetches all existing tasks in queue and check if those are duplicate with new tasks or not. For e.g. if 100s of thousand of put mapping tasks are in queue, for new single put mapping task, it will iterate over all existing task to check for duplicate.

Time complexity : O (# of tasks in queue)

#### New logic
Cluster manager will maintain new map as well to store the task identities as well. We will introduce `taskIdentityPerBatchingKey` map, which will store task identity of all existing tasks against batching key. Against one batching key we will store LinkedHashSet for storing task identity, so we can check if identity is already exists or not in O(1). Once new tasks are submitted, we will check if those new task identity is already present in existing identity or not. For e.g. if 100s of thousand of put mapping tasks are in queue, for new single put mapping task, it will check if task is exist in existing identity or not. 

Time complexity : O (# of new tasks getting submitted)

### Improvement

Used below test case for assessing performance.

https://github.com/opensearch-project/OpenSearch/blob/082f0594582116713355ae00fe5ad001774a511d/server/src/test/java/org/opensearch/cluster/service/TaskBatcherTests.java#L214

Ran this test with different setup (changed threads and tasksPerThread) and ran it 5 time and took avg of test execution time.

We have used 100 task executors[threads] in below experiments.

| Setup [Tasks Per executor] | Time taken without improvement [in s] | Time taken with improvement [in s] | Improvement [OldTime/NewTime] |
| ------------- | ------------- | ------------- | ------------- |
| 2500  | 8 | 2.5 | 320%|
| 5000  | 27 | 7 | 385% |
| 10000  | 83 | 22 | 377% |
| 20000  | 268 | 88 | 304% |

### Issues Resolved
#6567 

### Check List
- [ ] New functionality includes testing.
  - [ ] All tests pass
- [ ] New functionality has been documented.
  - [ ] New functionality has javadoc added
- [ ] Commits are signed per the DCO using --signoff
- [ ] Commit changes are listed out in CHANGELOG.md file (See: [Changelog](../blob/main/CONTRIBUTING.md#changelog))

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).